chore: move alt constructor back to model
